Docs: Use main branch links to source code
The main branch evolves, and the documentation should evolve with it.
Link to TypeScript interfaces and code samples on the main branch to
avoid confusing users with outdated information.

The documentation can still go out-of-date, of course. Mitigate that a
bit by avoiding to link exact line numbers within a file, since they are
most likely to change.