Spawn processes with LANG that's guaranteed to exist
`LANG=en_US` is not guaranteed to exist on any system (it does not, for
strange reasons, exist on mine). Choose `LANG=C`, which is guaranteed to
exist. It has the side-effect of typically using an encoding other than
UTF-8 for the tool output (e.g., ISO 8859-1). Apart from messages
printed into the log in this encoding that should hopefully have no
negative side-effects.

See https://github.com/sorenlouv/backport/pull/505 for a case where
code was assuming that output would always be English, and failed if it
wasn't.