Allow configuring RegEx patterns for `copySourcePRLabels`
Currently, we only allow copying over either all or no labels. With this
commit, we also allow specifying a list of RegEx patterns that will be
used to conditionally copy over labels.

This is useful when you don't want to copy over all labels, but you also
don't want to manually specify every time which labels you want to be
copied over.